Create rsa_keygen.java
I put the functions in the class because I didn't think they were usable by the other files.  If we decide to create a function file, I can move them.  

It creates public and private keys, as well as a short encrypt and decrypt of an integer (we can take out later), to make sure it works .

I didn't know what he was talking about with the "r should be half the bit size of N" , so something might have to change.